conformance/buffimgs/cl_commandQueue_enqueueCopyBufferRect.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueCopyBufferRect conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>

try {
    debug("WebCLCommandQueue enqueueCopyBufferRect test case. \n");
    debug("( WebCLBuffer srcBuffer, WebCLBuffer dstBuffer, CLuint[3] srcOrigin, CLuint[3] dstOrigin, CLuint[3] region, CLuint srcRowPitch, CLuint srcSlicePitch, CLuint dstRowPitch, CLuint dstSlicePitch, optional WebCLEvent[]? eventWaitList, optional WebCLEvent? event)");

    var WIDTH_STEP = Float32Array.BYTES_PER_ELEMENT * 320 * 4;
    var BUFFER_SIZE = Float32Array.BYTES_PER_ELEMENT * 320 * 240 * 4;
    var srcRowPitch = WIDTH_STEP;
    var srcSlicePitch = WIDTH_STEP;
    var dstRowPitch = WIDTH_STEP;
    var dstSlicePitch = WIDTH_STEP;

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var srcBufferRead = webCLContext.createBuffer(webcl.MEM_READ_ONLY, BUFFER_SIZE);
    var srcBufferWrite = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, BUFFER_SIZE);
    var srcBufferReadWrite = webCLContext.createBuffer(webcl.MEM_READ_WRITE, BUFFER_SIZE);
    var srcOrigin = new Int32Array([12, 25, 0]);
    var dstOrigin = new Int32Array([10, 50, 0]);
    var region = new Int32Array([10, 20, 1]);

    // Testing with destination buffer  as MEM_READ_WRITE  memory object.
    var dstBufferReadWrite = webCLContext.createBuffer(webcl.MEM_READ_WRITE, BUFFER_SIZE);
    shouldBeTrue("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBufferReadWrite, dstBufferReadWrite, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);");
    shouldBeTrue("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBufferRead, dstBufferReadWrite, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBufferWrite, dstBufferReadWrite, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");

    // Testing with destination buffer  as MEM_READ_ONLY memory object.
    var dstBufferRead = webCLContext.createBuffer(webcl.MEM_READ_ONLY, BUFFER_SIZE);
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBufferReadWrite, dstBufferRead, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBufferRead, dstBufferRead, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBufferWrite, dstBufferRead, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");

    // Testing with destination buffer  as MEM_WRITE_ONLY memory object.
    var dstBufferWrite = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, BUFFER_SIZE);
    shouldBeTrue("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBufferReadWrite, dstBufferWrite, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);");
    shouldBeTrue("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBufferRead, dstBufferWrite, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBufferWrite, dstBufferWrite, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");

    var srcBuffer = webCLContext.createBuffer(webcl.MEM_READ_WRITE, BUFFER_SIZE);
    var dstBuffer = webCLContext.createBuffer(webcl.MEM_READ_WRITE, BUFFER_SIZE);

    // Testing for invalid values of Buffer Origin.
    var srcOriginInvalid = new Int32Array([-12, -25, 0]);
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOriginInvalid, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");
    var dstOriginInvalid = new Int32Array([-10, -50, 0]);
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOriginInvalid, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");

    // Testing for values of rectangular region.
    var regionInvalid = new Int32Array([-10, -20]);
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, regionInvalid, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");

    // Testing for values of Source row pitch and Source slice pitch
    var srcRowPitchInvalid = 200;
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, region, srcRowPitchInvalid, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");
    var srcSlicePitchInvalid = 200;
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitchInvalid, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");

    // Testing for values of Destination row pitch and Destination slice pitch
    var dstRowPitchInvalid = 200;
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitchInvalid, dstSlicePitch);", "INVALID_VALUE");
    var dstSlicePitchInvalid = 200;
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitchInvalid);", "INVALID_VALUE");

    // Testing invalid objects as params.
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, invalid_object, dstBuffer, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_MEM_OBJECT");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, invalid_object, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_MEM_OBJECT");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, invalid_object, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, invalid_object, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, invalid_object, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, region, invalid_object, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, region, srcRowPitch, invalid_object, dstRowPitch, dstSlicePitch);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, invalid_object, dstSlicePitch);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandQueue, srcBuffer, dstBuffer, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, invalid_object);", "INVALID_VALUE");

    // FIXME :: Need to add test cases passing event parameters.

    // Testing for exception "INVALID_CONTEXT".
    var webCLContext2 = wtu.createContext();
    var webCLCommandqueue2 = wtu.createCommandQueue(webCLContext2);
    shouldThrowExceptionName("wtu.enqueueCopyBufferRect(webCLCommandqueue2, srcBuffer, dstBuffer, srcOrigin, dstOrigin, region, srcRowPitch, srcSlicePitch, dstRowPitch, dstSlicePitch);", "INVALID_CONTEXT");

} catch (e) {
    testFailed("WebCLCommandQueue :: enqueueCopyBufferRect throws an exception : " + e.message);
}
notifyFinishedToHarness();
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
